refactor: 모듈화(#7)
